requirements.txt
Django==2.2.4
django-oauth-toolkit
django-rest-framework